Remove p7000-filter-branch
This test existed to measure performance of filter-branch, having been
used twice in history.  However, the performance framework brings a lot
of overhead with it, anyone interested in performance should just not
use filter-branch as filter-repo is a hundred times faster (even if they
absolutely *need* the filter-branch user interface for some odd reason,
they should use the reimplementation of it provided as a contrib script
in filter-repo which is several times faster than filter-branch), and it
is unlikely filter-branch will be seeing much future development anyway.
As such, the overhead of the performance testing in git.git was not
worth copying over, so just remove this test.

Signed-off-by: Elijah Newren <newren@gmail.com>