Adjust git-filter-branch for living within its own repo
We no longer get to assume that git-sh-setup is in the same directory,
so instead we query $(git --exec-path) for that information.  Further,
git.git had an involved cmd_munge_script for translating many variables,
and had a lot of code around creating and updating this script based on
lots of different input.  Keeping all of that setup just to translate
'/bin/sh' to something else seems to be overkill.  As such, we can just
have the documentation mention this fix for the oddball platforms that
need it.

Signed-off-by: Elijah Newren <newren@gmail.com>