Streamline build files
This repo doesn't have all the same files that git.git had; it exists
just for git-filter-branch and as such the Makefiles can be simplified
pretty dramatically.  Also, the test framework needs to avoid checking
for a built version of 'git' in this project and instead rely on an
external installation.

Signed-off-by: Elijah Newren <newren@gmail.com>